// scripts/atlas-json/hierarchical/powerhouse-types/types/view-nodes.ts
export type RawViewNode = {
  id: string;
  type: TDocType;
  title: ViewNodeTitle;
  content: TNodeContent;
  rawContent: RichTextItemResponse[];
  slugSuffix: string;
  parentSlugSuffix: string | null;
  ancestorSlugSuffixes: string[];
  descendantSlugSuffixes: string[];
  subDocuments: RawViewNode[];
  globalTags: string[];
  originalContextData: string[];
  masterStatus: string | null;
  files: TProcessedFile[];
};

/**
 * Extended view node type with additional properties
 * - markdownContent: Markdown content string
 * - globalTags: Global tags
 */
export type ViewNodeExtended = Omit<RawViewNode, 'content' | 'subDocuments' | 'rawContent'> & {
  content: TProcessedNodeContentItem[];
  subDocuments: ViewNodeExtended[];
  markdownContent: string;
};
// ...
export type ViewNode = Omit<
  ViewNodeExtended,
  'subDocuments' | 'globalTags' | 'originalContextData' | 'markdownContent' | 'masterStatus'
> & {
  subDocuments: ViewNode[];
};
// ...
export type TProcessedNodeContentItem = {
  heading?: string | null;
  text: TProcessedViewNodeContent[];
};